+++ noodles-vcf/CHANGELOG.md
@@ -41,6 +41,8 @@
 
 ### Changed
 
+  * vcf/header/parser: Return error on duplicate info ID.
+
   * vcf/reader/builder: Change `Builder::build_from_reader` to accept
     implementations of `Read`.